fix(groups): patched JS/TS `const` coloring to be consistent
Fixes an issue in JS/TS with `const` variables *always* being colored as a constant, even if it contains an anonymous function.

This patch disables LSP highlighting of variables and constants, so that this job is delegated to Tree-sitter, which highlights depending on variable's casing. Note that LSP's function highlighting is still active, so the coloring still carries some type information.